# docs/api.yml
openapi: 3.0.0
info:
  title: Reframe API
  description: Reframe Knowledge Management API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 開発環境
  - url: https://api.reframe.example.com
    description: 本番環境
...
paths:
...
  /api/knowledge:
    get:
      summary: 全ナレッジ取得
      description: 登録されているすべてのナレッジを取得します
      tags:
        - Knowledge
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Knowledge'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: ナレッジ作成
      description: 新しいナレッジを作成します
      tags:
        - Knowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKnowledgeInput'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Knowledge'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Knowledge:
      type: object
      required:
        - id
        - title
        - content
        - embedding
        - connections
        - createdAt
      properties:
        id:
          type: string
          description: ナレッジのID
        title:
          type: string
          description: ナレッジのタイトル
        content:
          type: string
          description: ナレッジの内容
        embedding:
          type: array
          items:
            type: number
          description: 埋め込みベクトル
        connections:
          type: array
          items:
            type: string
          description: 接続されているナレッジのIDリスト
        createdAt:
          type: string
          format: date-time
          description: 作成日時
    CreateKnowledgeInput:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          description: ナレッジのタイトル
        content:
          type: string
          description: ナレッジの内容
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: エラーメッセージ